+++ src/test/java/com/mageddo/dnsproxyserver/AppIntTest.java
@@ -4,10 +4,12 @@ import com.mageddo.commons.concurrent.Threads;
 import com.mageddo.dns.utils.Messages;
 import com.mageddo.dnsproxyserver.solver.SimpleResolver;
 import com.mageddo.dnsproxyserver.utils.Ips;
-import com.mageddo.net.SocketUtils;
 import com.mageddo.utils.Executors;
+import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
+import org.xbill.DNS.Message;
+import testing.templates.ConfigFlagArgsTemplates;
 
 import java.io.IOException;
 import java.time.Duration;
@@ -19,32 +21,29 @@ public class AppIntTest {
   @Test
   void appMustStart() throws IOException {
 
-    final var webServerPort = SocketUtils.findRandomFreePort();
-    final var dnsServerPort = SocketUtils.findRandomFreePort();
     final var hostToQuery = "dps-sample.dev";
-
-    final var config = new String[]{
-      "--default-dns=false",
-      "--web-server-port=" + webServerPort,
-      "--server-port=" + dnsServerPort,
-      "--log-level=TRACE",
-    };
-    final var app = new App(config);
+    final var args = ConfigFlagArgsTemplates.withRandomPortsAndNotAsDefaultDns();
+    final var app = new App(args);
 
     try (final var executor = Executors.newThreadExecutor()) {
+
       executor.submit(app::start);
 
       Threads.sleep(Duration.ofSeconds(2));
 
-      final var dnsServerAddress = Ips.getAnyLocalAddress(dnsServerPort);
-
-      final var dnsClient = new SimpleResolver(dnsServerAddress);
-      final var res = dnsClient.send(Messages.aQuestion(hostToQuery));
-
+      final var port = app.getDnsServerPort();
+      final var res = queryStartedServer(port, hostToQuery);
       assertTrue(Messages.isSuccess(res));
 
     }
 
+  }
 
+  @SneakyThrows
+  static Message queryStartedServer(Integer port, String host) {
+    final var dnsServerAddress = Ips.getAnyLocalAddress(port);
+    final var dnsClient = new SimpleResolver(dnsServerAddress);
+    final var res = dnsClient.send(Messages.aQuestion(host));
+    return res;
   }
 }
